scheduler: Some fixes to different lr decay options

+++ scripts/train_vilbert.py
@@ -521,7 +521,7 @@ def main():
     parser.add_argument("--no-lr-decay", action="store_true")
     parser.add_argument("--wd", type=float, default=1e-2)
     parser.add_argument("--warmup-proportion", type=float, default=0.1)
-    parser.add_argument("--decay-power", type=int, default=-1)
+    parser.add_argument("--decay-power", type=float, default=-1)
     parser.add_argument("--iterations", type=int, default=2500000)
     parser.add_argument("--disable-shuffle", action="store_true")
     parser.add_argument("--check-val-every", type=int, default=1000)
